% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{fit_statistics}
\alias{fit_statistics}
\alias{fit_statistics.lc_mpt}
\alias{fit_statistics.lc_mpt_list}
\title{Fit Statistics for Latent-class MPT models}
\usage{
fit_statistics(x, ...)

\method{fit_statistics}{lc_mpt}(x, ...)

\method{fit_statistics}{lc_mpt_list}(x, ...)
}
\description{
Extracts fit statistics from latent-class MPT model objects.
}
