% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lc.R
\name{lc}
\alias{lc}
\title{Fit latent-class MPT models}
\usage{
lc(model_file, data_file, classes = NULL, max_classes = 20,
  runs = 20, fisher_information = "expected",
  montecarlo_samples = 1e+05, crit = "AIC", verbose = FALSE)
}
\arguments{
\item{classes, max_classes}{The number of classes for a single model or the
maximum number of classes. To estimate a single model, specify \code{classes};
to estimate multiple models with increasing number of classes, specify \code{max_classes}.}

\item{fisher_information}{The type of Fisher Information to be computed.
Can be either \code{expected}, \code{montecarlo}, \code{observed}, or \code{none}.
Defaults to \code{expected}. However, if expected Fisher Information cannot be
computed, the Monte Carlo method is used.}

\item{montecarlo_samples}{The number of simulations to be used for Monte Carlo
Fisher Information. Defaults to 1e5.}

\item{verbose}{Logical. Indicating whether an announcement is printed on the
console when fitting of a new model starts.}
}
\description{
This function estimates one or more latent-class MPT models.
Starting from a one-class solution, it increments the number of latent classes
until a criterion is met.
}
\references{
Stahl, C., & Klauer, K.C. (2007). HMMTree: A computer program for hierarchical
  multinomial processing tree models. \emph{Behavior Research Methods}, \emph{39}, 267-273.
}
