% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmmtree.R
\name{lc}
\alias{lc}
\title{Fit latent-class MPT models}
\usage{
lc(model, data, nsubj, nclass_max = 5, nruns = 1, fi = 3, mc = 10000,
  comma = 0, path = "./", crit = "AIC")
}
\description{
This function estimates one or more latent-class MPT models.
Starting from a one-class solution, it increments the number of latent classes
until a criterion is met.
}
\references{
Stahl, C., & Klauer, K.C. (2007). HMMTree: A computer program for hierarchical multinomial processing tree models. \emph{Behavior Research Methods}, \emph{39}, 267-273.
}
