% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{parameter_estimates}
\alias{parameter_estimates}
\alias{parameter_estimates.lc_mpt}
\alias{parameter_estimates.lc_mpt_list}
\title{Parameter Estimates from Latent-class MPT models}
\usage{
parameter_estimates(x, ...)

\method{parameter_estimates}{lc_mpt}(x, ...)

\method{parameter_estimates}{lc_mpt_list}(x, ...)
}
\arguments{
\item{x}{An object of classes \code{lc_mpt} or \code{lc_mpt_list}.}

\item{...}{Further arguments that may be passed to methods, currently ignored.}
}
\description{
Extracts parameter estimates from latent-class MPT model objects.
}
