% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{weighted_means}
\alias{weighted_means}
\alias{weighted_means.lc_mpt}
\alias{weighted_means.lc_mpt_list}
\title{Weighted Means of Parameter Estimates}
\usage{
weighted_means(x, ...)

\method{weighted_means}{lc_mpt}(x, ...)

\method{weighted_means}{lc_mpt_list}(x, ...)
}
\arguments{
\item{x}{An object of class \code{lc_mpt} or \code{lc_mpt_list}.}
}
\description{
Extract weighted means of parameter estimates.
}
